fix: 누락된 sql문 수정
