+++ click-service/src/main/resources/db/migration/V1__init.sql
@@ -1,6 +1,7 @@
 create table member
 (
     id bigint       not null auto_increment,
+    click_count bigint not null,
     name     varchar(255) not null,
     profile_image_url  varchar(255),
     created_at datetime not null,
